no message
git-svn-id: https://svn.grrrr.org/ext/trunk@450 4d9ac71a-51e6-0310-8455-cad1006bcd31

+++ source/flsupport.h
@@ -155,7 +155,7 @@ public:
 	//! Zero a memory region
 	static void ZeroMem(void *dst,int bytes);
 	//! Sleep for an amount of time
-	static void Sleep(float s);
+	static void Sleep(double s);
 
 //!		@} FLEXT_S_UTIL
